Block Radius Changes
Changelog:
- Renamed Revoker Area feature to Block Radius
- Disruptors now also have a radius
- Block Radius doesn't disappear when going further than 16 blocks
- Updated currency formatting to #.###[suffix] to #.##[suffix]
- Fixed some scoreboard coloring issues

+++ src/main/java/com/github/vaapukkax/kuphack/flagclash/FlagClash.java
@@ -79,7 +79,7 @@ public class FlagClash implements EventHolder {
 	
 	public static String timeAsString(double time) {
 		if (time < 60) return (int) time + "s";
-		String text = time / 60 % 60 + "min";
+		String text = (int)(time / 60 % 60) + "min";
 		if (time > 3600) text = Kuphack.round(time / 60.0 / 60.0) + "h";
 		return text;
 	}
